Update/Opt 0.06
Big changes. Huge optimization to the hashspace search by rolling collision up a level and reducing the number of grid searches to the number of buckets rather than the number of objects in the world. Eliminated the simplespace/listworld as they were causing too many problems with inheritance and namespace pollution. Fixed bucket tracking on the grid. Started changing style from 4 spaces to 2 spaces per indentation, indentation is mixed currently.

In the process of adding line collisions and possibly removing world in favor of a an OO engine.